+++ Sources/Atomic.swift
@@ -226,6 +226,13 @@ public final class Atomic<Value> {
 		set(newValue) {
 			swap(newValue)
 		}
+
+		_modify {
+			lock.lock()
+			defer { lock.unlock() }
+
+			yield &_value
+		}
 	}
 
 	/// Initialize the variable with the given initial value.